+++ test/src/plugin/auto_detecting_codec_test.dart
@@ -1,4 +1,5 @@
 import 'package:rdf_core/rdf_core.dart';
+import 'package:rdf_core/src/plugin/rdf_codec_registry.dart';
 import 'package:test/test.dart';
 
 // Tests for the AutoDetectingRdfCodec and AutoDetectingRdfDecoder classes